requirements.txt
Django>=5.0.6,<5.1.0
djangorestframework>=3.15.1,<3.16.0
dj-database-url>=2.2.0,<2.3.0
pillow>=10.3.0,<10.4.0
psycopg2>=2.9.9,<2.10.0
requests>=2.32.3,<2.33.0
uwsgi>=2.0.20,<2.1